BE_Project/requirements.txt
# Core Dependencies for IntelliQuery

# Core Framework and Database
flask>=2.3.0
flask-cors>=4.0.0
mysql-connector-python>=8.0.0
qdrant-client>=1.1.1
sqlparse>=0.4.0

# Google Generative AI and ML
google-generativeai>=0.3.0

# Machine Learning and NLP
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0

# Data Processing and Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
pandas>=2.1.0

# Text Processing and Language Detection
langdetect>=1.0.9
dateparser>=1.1.8

# HTTP Requests and API
requests>=2.31.0
httpx>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: For enhanced functionality
# Uncomment if needed:
# openai>=1.3.0  # Alternative LLM provider
# anthropic>=0.7.0  # Alternative LLM provider
# redis>=5.0.0  # For caching
# celery>=5.3.0  # For async task processing
# sqlalchemy>=2.0.0  # For database operations
# alembic>=1.13.0  # For database migrations

# Note: After installing, run:
# python -m spacy download en_core_web_sm
# to download the English spaCy model


google-generativeai>=0.1.5
pydantic>=1.10
spacy>=3.5
python-dotenv>=1.0
dateparser>=1.1
pytest>=7.0
requests
